requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib
python-jose[cryptography]
python-multipart
bcrypt
pyjwt
redis
pytest
pytest-cov
httpx
locust
pytest-mock
codecov
